Refactor to use commons utils
