+++ maat-orchestration/build.gradle
@@ -20,7 +20,7 @@ def versions = [
 		pitest                  : "1.15.3",
 		sentryVersion			: "7.1.0",
 		springdocVersion        : "2.3.0",
-		crimeCommonsVersion     : "2.12.0",
+		crimeCommonsVersion     : "2.13.3",
 		okhttpVersion           : "4.12.0",
 		wiremockVersion    		: "3.0.1"
 ]
@@ -56,6 +56,7 @@ dependencies {
 	implementation "org.springframework.boot:spring-boot-starter-validation"
 
 	implementation "uk.gov.justice.service.laa-crime:crime-commons-spring-boot-starter-rest-client:$versions.crimeCommonsVersion"
+	implementation "uk.gov.justice.service.laa-crime:crime-commons-classes:$versions.crimeCommonsVersion"
 
 	compileOnly "org.projectlombok:lombok"
 	annotationProcessor "org.projectlombok:lombok"
