--- maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/annotation/DefaultHTTPErrorResponse.java
@@ -1,32 +1,0 @@
-package uk.gov.justice.laa.crime.orchestration.annotation;
-
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import org.springframework.http.MediaType;
-import uk.gov.justice.laa.crime.orchestration.dto.ErrorDTO;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-@ApiResponses(value = {
-        @ApiResponse(responseCode = "400",
-                description = "Bad Request.",
-                content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
-                        schema = @Schema(implementation = ErrorDTO.class)
-                )
-        ),
-        @ApiResponse(responseCode = "500",
-                description = "Server Error.",
-                content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
-                        schema = @Schema(implementation = ErrorDTO.class)
-                )
-        )
-})
-public @interface DefaultHTTPErrorResponse {
-}
