--- maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/dto/ErrorDTO.java
@@ -1,12 +1,0 @@
-package uk.gov.justice.laa.crime.orchestration.dto;
-
-import lombok.Builder;
-import lombok.Value;
-
-@Value
-@Builder
-public class ErrorDTO {
-    String traceId;
-    String code;
-    String message;
-}
