+++ maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/exception/DefaultExceptionHandler.java
@@ -6,9 +6,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import uk.gov.justice.laa.crime.commons.dto.ErrorDTO;
 import uk.gov.justice.laa.crime.commons.exception.APIClientException;
 import uk.gov.justice.laa.crime.commons.tracing.TraceIdHandler;
-import uk.gov.justice.laa.crime.orchestration.dto.ErrorDTO;
 
 @Slf4j
 @RestControllerAdvice
