+++ maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/mapper/CrownCourtMapper.java
@@ -5,7 +5,7 @@ import uk.gov.justice.laa.crime.orchestration.dto.maat.CrownCourtSummaryDTO;
 import uk.gov.justice.laa.crime.orchestration.dto.maat.OutcomeDTO;
 import uk.gov.justice.laa.crime.orchestration.enums.CrownCourtOutcome;
 import uk.gov.justice.laa.crime.orchestration.model.common.ApiCrownCourtOutcome;
-import uk.gov.justice.laa.crime.orchestration.util.DateUtil;
+import uk.gov.justice.laa.crime.util.DateUtil;
 
 import java.util.Collection;
 import java.util.List;
