--- maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/util/NumberUtils.java
@@ -1,16 +1,0 @@
-package uk.gov.justice.laa.crime.orchestration.util;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class NumberUtils {
-
-    public static Integer toInteger(Long value) {
-        return (value != null) ? value.intValue() : null;
-    }
-
-    public static Double toDouble(Integer input) {
-        return (input != null) ? input.doubleValue() : null;
-    }
-}
