+++ maat-orchestration/src/test/java/uk/gov/justice/laa/crime/orchestration/controller/MeansAssessmentControllerTest.java
@@ -24,8 +24,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static uk.gov.justice.laa.crime.orchestration.data.Constants.APPLICANT_ID;
 import static uk.gov.justice.laa.crime.orchestration.data.Constants.FINANCIAL_ASSESSMENT_ID;
-import static uk.gov.justice.laa.crime.orchestration.util.RequestBuilderUtils.buildRequest;
-import static uk.gov.justice.laa.crime.orchestration.util.RequestBuilderUtils.buildRequestGivenContent;
+import static uk.gov.justice.laa.crime.util.RequestBuilderUtils.buildRequestWithTransactionId;
+import static uk.gov.justice.laa.crime.util.RequestBuilderUtils.buildRequestWithTransactionIdGivenContent;
 
 @WebMvcTest(MeansAssessmentController.class)
 @Import(OrchestrationTestConfiguration.class)
@@ -50,14 +50,14 @@ class MeansAssessmentControllerTest {
                 .thenReturn(new FinancialAssessmentDTO());
 
         String endpoint = ENDPOINT_URL + "/" + FINANCIAL_ASSESSMENT_ID + "/applicantId/" + APPLICANT_ID;
-        mvc.perform(buildRequest(HttpMethod.GET, endpoint))
+        mvc.perform(buildRequestWithTransactionId(HttpMethod.GET, endpoint, true))
                 .andExpect(status().isOk())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON));
     }
 
     @Test
     void givenInvalidRequest_whenFindIsInvoked_thenBadRequestResponseIsReturned() throws Exception {
-        mvc.perform(buildRequest(HttpMethod.GET, ENDPOINT_URL + "/invalidId" + "/applicantId/" + APPLICANT_ID))
+        mvc.perform(buildRequestWithTransactionId(HttpMethod.GET, ENDPOINT_URL + "/invalidId" + "/applicantId/" + APPLICANT_ID, true))
                 .andExpect(status().isBadRequest());
     }
 
@@ -67,7 +67,7 @@ class MeansAssessmentControllerTest {
                 .thenThrow(new APIClientException());
 
         String endpoint = ENDPOINT_URL + "/" + FINANCIAL_ASSESSMENT_ID + "/applicantId/" + APPLICANT_ID;
-        mvc.perform(buildRequest(HttpMethod.GET, endpoint))
+        mvc.perform(buildRequestWithTransactionId(HttpMethod.GET, endpoint, true))
                 .andExpect(status().isInternalServerError());
     }
 
@@ -84,14 +84,14 @@ class MeansAssessmentControllerTest {
         when(orchestrationService.create(any(WorkflowRequest.class)))
                 .thenReturn(new ApplicationDTO());
 
-        mvc.perform(buildRequestGivenContent(HttpMethod.POST, requestBody, ENDPOINT_URL))
+        mvc.perform(buildRequestWithTransactionIdGivenContent(HttpMethod.POST, requestBody, ENDPOINT_URL, true))
                 .andExpect(status().isOk())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON));
     }
 
     @Test
     void givenInvalidRequest_whenCreateIsInvoked_thenBadRequestResponseIsReturned() throws Exception {
-        mvc.perform(buildRequestGivenContent(HttpMethod.POST, "", ENDPOINT_URL))
+        mvc.perform(buildRequestWithTransactionIdGivenContent(HttpMethod.POST, "", ENDPOINT_URL, true))
                 .andExpect(status().isBadRequest());
     }
 
@@ -108,7 +108,7 @@ class MeansAssessmentControllerTest {
         when(orchestrationService.create(any(WorkflowRequest.class)))
                 .thenThrow(new APIClientException());
 
-        mvc.perform(buildRequestGivenContent(HttpMethod.POST, requestBody, ENDPOINT_URL))
+        mvc.perform(buildRequestWithTransactionIdGivenContent(HttpMethod.POST, requestBody, ENDPOINT_URL, true))
                 .andExpect(status().isInternalServerError())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON));
     }
@@ -124,14 +124,14 @@ class MeansAssessmentControllerTest {
         when(orchestrationService.update(any(WorkflowRequest.class)))
                 .thenReturn(new ApplicationDTO());
 
-        mvc.perform(buildRequestGivenContent(HttpMethod.PUT, requestBody, ENDPOINT_URL))
+        mvc.perform(buildRequestWithTransactionIdGivenContent(HttpMethod.PUT, requestBody, ENDPOINT_URL, true))
                 .andExpect(status().isOk())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON));
     }
 
     @Test
     void givenInvalidRequest_whenUpdateIsInvoked_thenBadRequestResponseIsReturned() throws Exception {
-        mvc.perform(buildRequestGivenContent(HttpMethod.PUT, "requestBody", ENDPOINT_URL))
+        mvc.perform(buildRequestWithTransactionIdGivenContent(HttpMethod.PUT, "requestBody", ENDPOINT_URL, true))
                 .andExpect(status().isBadRequest());
     }
 
@@ -146,8 +146,8 @@ class MeansAssessmentControllerTest {
         when(orchestrationService.update(any(WorkflowRequest.class)))
                 .thenThrow(new APIClientException());
 
-        mvc.perform(buildRequestGivenContent(HttpMethod.PUT, requestBody, ENDPOINT_URL))
-                .andExpect(status().isInternalServerError())
+        mvc.perform(buildRequestWithTransactionIdGivenContent(HttpMethod.PUT, requestBody, ENDPOINT_URL, true))
+                .andExpect(status().is5xxServerError())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON));
     }
